TXT Activity CSS
Build a HTML web page with the following:

2 Divs
Each div must be 500px in height
The top div must have a video playing and header element layered on top
The bottom div must have a solid background color with 2 paragraph elements
All CSS must be added via external css stylesheet

Note: My overlap will not save, so the text will not be over the video. However, I have a screenshot that presents that I was able to do it, however it will not save and I had trouble as to why not.